refactor(test): split fixtures into runtime, http, db, and operator packages
- move runtime bootstrap and lifecycle management into internal/test/runtime with a thin public facade in internal/test/test.go
- split HTTP helpers into dedicated fixture files for handler, CSRF, bearer token, and request utilities
- move migration fixtures into internal/test/fixture/db and export explicit RDBMS and ClickHouse fixture types
- move user and permission test helpers into internal/test/fixture/operator and wire them through the runtime composition root
- update controller and registrar test suites to use HTTP fixture methods instead of package-level request helper functions
- align README test structure and examples with the current runtime and fixture layout
